Merge to test

+++ NMP-Portal/NMP-Portal/ServiceResponses/NitrogenApplicationsForNMaxReportResponse.cs
@@ -9,9 +9,9 @@ namespace NMP.Portal.ServiceResponses
         public string CropTypeName { get; set; } = string.Empty;
         public decimal CropArea { get; set; }
         public decimal? InorganicNRate { get; set; }
-        public decimal? InorganicNTotal { get; set; }
+        public int? InorganicNTotal { get; set; }
         public decimal? OrganicCropAvailableNRate { get; set; }
-        public decimal? OrganicCropAvailableNTotal { get; set; }
+        public int? OrganicCropAvailableNTotal { get; set; }
         public decimal? NRate { get; set; }
         public decimal? NTotal { get; set; }
     }
